Replace `WebAuthSession` with `WebAuthState`, refactor authentication flow, and update dependencies.
- Introduced `WebAuthState` to unify and simplify authentication state management across platforms.
- Replaced `WebAuthSession` usage in `Main.kt` for both browser and WASM.
- Updated `build.gradle.kts` to include new dependency `kotlinx-browser`.
- Adjusted `API_BASE_URL` for local debugging.
