+++ composeApp/src/jsMain/kotlin/org/jetbrains/Main.kt
@@ -3,36 +3,20 @@ package org.jetbrains
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import androidx.navigation.ExperimentalBrowserHistoryApi
-import androidx.navigation.bindToNavigation
-import kotlinx.browser.window
+import androidx.navigation.NavController
+import androidx.navigation.bindToBrowserNavigation
+import org.jetbrains.demo.config.WebAuthState
 import org.jetbrains.demo.di.appModule
 import org.jetbrains.demo.ui.App
-import org.jetbrains.demo.ui.AuthSession
 import org.jetbrains.skiko.wasm.onWasmReady
 import org.koin.core.context.startKoin
-
-object WebAuthSession : AuthSession {
-    override fun hasToken(): Boolean {
-        val currentUrl = window.location.href
-        return when {
-            currentUrl.contains("login") -> false
-            else -> true
-        }
-    }
-
-    override fun clearToken() {
-        window.location.href = "/logout"
-    }
-}
 
 @OptIn(ExperimentalComposeUiApi::class, ExperimentalBrowserHistoryApi::class)
 fun main() {
     onWasmReady {
         startKoin { modules(appModule) }
         ComposeViewport("ComposeApp") {
-            App({ controller ->
-                window.bindToNavigation(controller)
-            }, WebAuthSession)
+            App(NavController::bindToBrowserNavigation, WebAuthState)
         }
     }
 }
