+++ composeApp/src/wasmJsMain/kotlin/org/jetbrains/Main.kt
@@ -1,30 +1,16 @@
 package org.jetbrains
 
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import androidx.navigation.ExperimentalBrowserHistoryApi
+import androidx.navigation.bindToBrowserNavigation
 import androidx.navigation.bindToNavigation
-import kotlinx.browser.document
 import kotlinx.browser.window
+import org.jetbrains.demo.config.WebAuthState
 import org.jetbrains.demo.di.appModule
 import org.jetbrains.demo.ui.App
-import org.jetbrains.demo.ui.AuthSession
+import org.jetbrains.demo.ui.AuthState
 import org.koin.core.context.startKoin
-
-object WebAuthSession : AuthSession {
-    override fun hasToken(): Boolean {
-        val currentUrl = window.location.href
-        return when {
-            currentUrl.contains("login") -> false
-            else -> true
-        }
-    }
-
-    override fun clearToken() {
-        window.location.href = "/logout"
-    }
-}
 
 @OptIn(
     ExperimentalComposeUiApi::class,
@@ -34,13 +20,13 @@ fun main() {
     startKoin { modules(appModule) }
     ComposeViewport("ComposeApp") {
         App({ controller ->
-            window.bindToNavigation(controller) { entry ->
+            controller.bindToBrowserNavigation { entry ->
                 val route = entry.destination.route.orEmpty()
                 when {
                     route == "chat" -> ""
                     else -> route
                 }
             }
-        }, WebAuthSession)
+        }, WebAuthState)
     }
 }
